+++ libs/guiutils/KoResourceChooser.cpp
@@ -50,7 +50,7 @@ QTextDocument *KoIconToolTip::createDocument( const QModelIndex &index )
     const QString html = QString( "<html><body>%1</body></html>" ).arg( body );
 
     doc->setHtml( html );
-    doc->setTextWidth( qMin( doc->size().width(), 500.0 ) );
+    doc->setTextWidth( qMin( doc->size().width(), qreal(500.0) ) );
 
     return doc;
 }
